Realized that individuals' phenotypes were being recalculated each timestep, so created Individual.phenotype attributes that are set once when each individual is created (and reset if genomes are reset after burn-in). Also created the Individual.idx attribute, since it was simple to do and made sense, and Jason and I talked about it a little while ago, and also made sure it is set everywhere where a new individual is created. Also implemented the new movement algorithm. (Note that the lookup array that serves as a set of approximations to the cell-wise Von Mises mixture distributions takes a long time to create up front for a reasonable number of histogram draws, e.g. 5000?? (need to write some tests to see how many really is 'reasonable'), but params['movement_surf_approx_len'] controls this, and can be kept set to really low values for quick work on the command line; and from what I've seen when testing with %timeit, runtime of r.choice on a numpy array is fairly insensitive to array size, so scaling up to a reasonable approx_len shouldn't heavily impact runtime after the up-front creation cost is sunk.) For more detail, see ./scripts_for_practice/profiling_and_optimizing/NOTES.txt.

+++ python_model/scripts_for_practice/SCRATCH_optimize_movement_and_recomb_algs.py
@@ -288,7 +288,7 @@ def alt_alt_create_movement_surface(land, params, kappa=12, approx_len=500):
     cells = landscape.get_cell_strings(gi, gj, max([len(str(land.dims[0])), len(str(land.dims[1]))]))
     cells = cells[::-1]
 
-    movement_surf = np.zeros((land.dims[0], land.dims[1], approx_len))
+    movement_surf = np.float16(np.zeros((land.dims[0], land.dims[1], approx_len)))
 
     for i in range(land.dims[0]):
         for j in range(land.dims[1]):
